requirements.txt
aiogram>=3.7,<4.0
yt-dlp[default]>=2024.4.9
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pyzipper>=0.3.6
aiohttp>=3.11.0
aiofiles>=24.1.0
feedparser>=6.0.11
telethon>=1.36.0
ruff>=0.4.0
pytest>=8.0
pytest-asyncio>=0.23
